[REFACTOR] Rename TextSelection to OldTextSelection
Renames TextSelection to OldTextSelection in preparation of the upcoming
refactoring introducing a new TextSelection implementation that works
based on line numbers and in-line offsets instead of text offsets.

The current implementation, now named OldTextSelection, will be removed
as part of the refactoring.

Subsequently rename EditorAPI.getSelection(IEditorPart) to
getOldSelection(...). This was done to avoid clashes with the new
implementation using the same signature.