[INTERNAL][I] Introduce BackgroundEditorPool
Adds BackgroundEditorPool to provide access to editor objects for
resources that are not currently open in an editor.

This will be necessary for the upcoming refactoring moving to a line
number and in-line offset based system to described positions in a text
document.  As the local logic still operates on text offsets, we have to
convert between the new positions and text offsets. Editors offer an
easy API to do such conversions. Furthermore, the API works on a caching
model, which should speed up the operations substantially.

The pool is backed by an LRU cache to avoid holding to many editor
references that are potentially no longer necessary, thereby creating a
quasi memory leak during a session. Instead, only the last 20 requested
editors are held and swapped out as necessary.

Adds the necessary methods to instantiate and release such editors to
the ProjectAPI.

Integrates the BackgroundEditorPool with the rest of the Saros logic to
clean it up according to state changes:
 - drop background editor when a real one is opened by the user
 - drop background editor when resource is deleted
 - drop background editor when session ends
 - drop background editor before application is torn down (necessary to
   avoid IntelliJ memory leak detection)

Adjusts the IntelliJ context test setup to mock the necessary IntelliJ
components used by the BackgroundEditorPool.