[INTERNAL] Add API calls to convert positioning schemes
Adds the necessary calls to the Saros/E and Saros/I EditorAPI to convert
between the text offset based and the line number and in-line offset
based positioning scheme.

These methods will be used in the upcoming refactoring migrating to the
line number and in-line offset based positioning scheme. The conversions
are vital for the refactoring as the rest of the Saros logic still works
on text offset based position data.

The logic introduced it backed by methods of the corresponding IDEs.
Hopefully, these methods are optimized internally to allow for a better
runtime.