[INTERNAL] Adjust logic in EditorAPI to handle single positions
Adjusts the text position handling in EditorAPI to offer methods to
convert single positions/offsets.

Refactors the existing logic to use the newly added single
position/offset logic to avoid code duplication.

To keep the performance optimization that can be gained when converting
values that are located in the same line, there is an internal version
of  the single offset conversion methods that also accepts and returns
the needed values for such optimizations.